removed redundant code and removed compareImage function to a test-util file
