+++ test/node-examples.js
@@ -25,21 +25,16 @@ const puppeteer = require('puppeteer');
 const console = require('console');
 const process = require('process');
 
-const fs = require('fs');
+// const fs = require('fs');
 const path = require('path');
-const PNG = require('pngjs').PNG;
-const pixelmatch = require('pixelmatch');
+// const PNG = require('pngjs').PNG;
+// const pixelmatch = require('pixelmatch');
+const compareImage = require('./compare-image');
 
 const LIB_DIR = path.resolve(__dirname, '..');
 const EXAMPLES_DIR = path.resolve(LIB_DIR, 'examples');
 
 let exampleDir;
-
-function printResult(diffRatio, threshold) {
-  return diffRatio <= threshold
-    ? console.log('\x1b[32m%s\x1b[0m', 'Rendering test Passed!')
-    : console.log('\x1b[31m%s\x1b[0m', 'Rendering test failed!');
-}
 
 async function validateWithWaitingTime(child, folder, waitingTime, threshold) {
   const browser = await puppeteer.launch({
@@ -53,40 +48,17 @@ async function validateWithWaitingTime(child, folder, waitingTime, threshold) {
   await page.waitFor(waitingTime);
   await page.screenshot({path: 'new.png'});
 
-  const goldImageData = fs.readFileSync(
-    path.resolve(
-      LIB_DIR,
-      'test/render/golden-images/examples/',
-      `${folder.replace(/\//g, '_')}.png`
-    )
-  );
-  const goldImage = PNG.sync.read(goldImageData);
-  const newImageData = fs.readFileSync('new.png');
-  const newImage = PNG.sync.read(newImageData);
-  const diffImage = new PNG({width: goldImage.width, height: goldImage.height});
-  const pixelDiffSize = pixelmatch(
-    goldImage.data,
-    newImage.data,
-    diffImage.data,
-    goldImage.width,
-    goldImage.height,
-    {threshold: 0.105, includeAA: true}
-  );
-  const pixelDiffRatio = pixelDiffSize / (goldImage.width * goldImage.height);
-
-  console.log(`Mismatched pixel number: ${pixelDiffSize}`);
-  console.log(`Mismatched pixel ratio: ${pixelDiffRatio}`);
-
-  const diffImageData = PNG.sync.write(diffImage, {colorType: 6});
-  fs.writeFileSync('diff.png', diffImageData);
-  fs.unlinkSync('new.png');
-  fs.unlinkSync('diff.png');
+  const golderImageName = `${LIB_DIR}/test/render/golden-images/examples/${folder.replace(
+    /\//g,
+    '_'
+  )}.png`;
+  const result = compareImage('new.png', golderImageName, threshold);
 
   child.kill();
   await page.waitFor(1000);
   await browser.close();
-  printResult(pixelDiffRatio, threshold);
-  return pixelDiffRatio <= threshold;
+
+  return result;
 }
 
 async function yarnAndLaunchWebpack() {
@@ -135,25 +107,25 @@ async function runTestExample(folder) {
   checkMapboxToken();
 
   exampleDir = EXAMPLES_DIR;
-  await runTestExample('experimental/bezier');
-  await runTestExample('experimental/json-pure-js');
-
-  await runTestExample('get-started/pure-js');
-  await runTestExample('get-started/pure-js-without-map');
-  await runTestExample('get-started/react-webpack-2');
-  await runTestExample('get-started/react-without-map');
-
-  await runTestExample('layer-browser');
-
-  await runTestExample('website/3d-heatmap');
-  await runTestExample('website/arc');
-  await runTestExample('website/brushing');
+  // await runTestExample('experimental/bezier');
+  // await runTestExample('experimental/json-pure-js');
+  //
+  // await runTestExample('get-started/pure-js');
+  // await runTestExample('get-started/pure-js-without-map');
+  // await runTestExample('get-started/react-webpack-2');
+  // await runTestExample('get-started/react-without-map');
+  //
+  // await runTestExample('layer-browser');
+  //
+  // await runTestExample('website/3d-heatmap');
+  // await runTestExample('website/arc');
+  // await runTestExample('website/brushing');
   await runTestExample('website/geojson');
-  await runTestExample('website/highway');
-  await runTestExample('website/icon');
-  await runTestExample('website/line');
-  await runTestExample('website/plot');
-  await runTestExample('website/scatterplot');
-  await runTestExample('website/screen-grid');
-  await runTestExample('website/tagmap');
+  // await runTestExample('website/highway');
+  // await runTestExample('website/icon');
+  // await runTestExample('website/line');
+  // await runTestExample('website/plot');
+  // await runTestExample('website/scatterplot');
+  // await runTestExample('website/screen-grid');
+  // await runTestExample('website/tagmap');
 })();
