+++ test/webpack.config.js
@@ -109,20 +109,6 @@ const CONFIGS = {
       entry: {
         app: resolve('./app.js')
       },
-      module: {
-        rules: [
-          {
-            test: /\.js$/,
-            exclude: [/node_modules/],
-            use: {
-              loader: 'babel-loader',
-              options: {
-                presets: ['@babel/preset-env']
-              }
-            }
-          }
-        ]
-      },
       plugins: [new HtmlWebpackPlugin()]
     }),
